Library

All functions exported by this package are listed below.

Library Contents

Library

Library Contents

Index

Functions

Plume Quantities

Unstratified Lazy Plume Equations

Stratified MTT Equations

Misc

Index

Plumes.BuoyancyFluxPerArea

Plumes.CalculateLinearDensityProfile

Plumes.CalculateMTTProfiles

Plumes.EntrainmentRate

Plumes.LSFitPlumeW

Plumes.LazyPlumeBuoyancy

Plumes.LazyPlumeFarFieldW

Plumes.LazyPlumeLengthScale

Plumes.LazyPlumeMomentumFlux

Plumes.LazyPlumeRadius

Plumes.LazyPlumeTracerConcentration

Plumes.LazyPlumeVolumeFlux

Plumes.LazyPlumeW

Plumes.LazyVirtualSourcePosition

Plumes.MeanCTDVariable

Plumes.MomentumFluxPerArea

Plumes.PlumeArea

Plumes.PlumeParameterΓ

Plumes.PlumeRe

Plumes.PlumeRi

Plumes.PurePlumeW

Plumes.ReducedGravity

Plumes.SourceRa

Plumes.TotalBuoyancyFlux

Plumes.TotalMomentumFlux

Plumes.TotalTracerFlux

Plumes.TotalVolumeFlux

Plumes.TracerFluxPerArea

Plumes.VolumeFluxPerArea

Plumes.mtt!

Plumes.mtt_fixedN²!

Plumes.mtt_free_α!

Plumes.mttmodel

Plumes.mttmodel_fixed_w₀

Plumes.mttmodel_free_α

Plumes.z2ζ

Plumes.Γ2ζ

Plumes.Γ_integral

Plumes.Γ₀Definition

Plumes.ζ2z

Plumes.ζ2Γ

Plumes.ρ_from_g_prime

Functions

Plume Quantities

Plumes.PlumeRe — Function

PlumeRe(w, b; ν=1.0e-6)

Calculate the plume Reynolds number from the plume velocity w, plume radius b, and kinematic viscosity ν.

Plumes.PlumeRi — Function

PlumeRi(w, b, g_prime)

Calculate the plume Richardson number from the plume velocity w, plume radius b, and reduced gravity g_prime.

PlumeRi(Γ; κ=2, α=0.11)

Calculate the source Richardson number from the source parameter Γ, shape coefficient κ, and entrainment coefficient α.

Plumes.SourceRa — Function

SourceRa(ρₐ, ρ₀, h; ν=1.0e-6, κ=1.0e-9, g = 9.80665)

Calculate the plume Rayleigh number from the ambient density ρₐ, source density ρ₀, depth h, kinematic viscosity ν, and molecular diffusivity κ.

SourceRa(g_prime₀, h; ν=1.0e-6, κ=1.0e-9)

Calculate the source Rayleigh number from the reduced gravity g_prime₀, depth h, kinematic viscosity ν, and molecular diffusivity κ.

Plumes.PlumeParameterΓ — Function

PlumeParameterΓ(w, b, g_prime; α=0.11)

Calculate the plume parameter Γ.

...

Arguments

b: Plume radius

w: Vertical velocity

g_prime: Reduced gravity

α: Entranment coefficient

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.Γ₀Definition — Function

Γ₀Definition(b₀, w₀, ρ₀, ρₐ; α=0.11, g = 9.80665)

Calculate the source plume parameter Γ₀.

...

Arguments

b₀: Source radius

w₀: Source vertical velocity

ρ₀: Souce density

ρₐ: Ambient density

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

α: Entranment coefficient

g: gravitational acceleration

...

Plumes.ReducedGravity — Function

ReducedGravity(ρ₀, ρₐ; g = 9.80665)

Calculate the reduced gravity for density ρ₀ and the ambient ρₐ.

Plumes.ρ_from_g_prime — Function

ρ_from_g_prime(g_prime, ρₐ; ρref=ρₐ, g = 9.80665)

Calculate the density ρ from the reduced gravity g_prime and the ambient density ρₐ.

Fluxes

Plumes.PlumeArea — Function

PlumeArea(b₀)

Area of the source of the plume with source radius b₀.

Plumes.TotalBuoyancyFlux — Function

TotalBuoyancyFlux(w₀, b₀, g_prime₀, κ)

Calculate the total buoyancy flux from the source velocity w₀, source radius b₀, and source reduced gravity g_prime₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.TotalMomentumFlux — Function

TotalMomentumFlux(w₀, b₀, κ)

Calculate the total momentum flux from the source velocity w₀, source radius b₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.TotalVolumeFlux — Function

TotalVolumeFlux(w₀, b₀)

Calculate the total volume flux from the source velocity w₀, source radius b₀. Volume flux does not depend on plume shape

Plumes.TotalTracerFlux — Function

TotalTracerFlux(w₀, b₀, κ; C₀=1)

Calculate the total tracer flux from the source velocity w₀, source radius b₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.BuoyancyFluxPerArea — Function

BuoyancyFluxPerArea(w₀, g_prime₀, κ)

Calculate the per area buoyancy flux from the source velocity w₀ and source reduced gravity g_prime₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.MomentumFluxPerArea — Function

MomentumFluxPerArea(w₀, κ)

Calculate the per area momentum flux from the source velocity w₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.VolumeFluxPerArea — Function

VolumeFluxPerArea(w₀)

Calculate the per area volume flux from the source velocity w₀. The volume flux does not depend on plume shape.

Plumes.TracerFluxPerArea — Function

TracerFluxPerArea(w₀, κ; C₀=1)

Calculate the per area tracer flux from the source velocity w₀ and source concentration C₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Unstratified Lazy Plume Equations

Plumes.z2ζ — Function

z2ζ(z, b₀; α=0.11)

Convert depth z to scaled plume height ζ using the source parameter b₀ and the entrainment coefficient α.

Plumes.ζ2z — Function

ζ2z(ζ, b₀; α=0.11)

Convert scaled plume height ζ to depth z using the source parameter b₀ and the entrainment coefficient α.

Plumes.Γ2ζ — Function

Γ2ζ(Γ, Γ₀)

Calculate the height parameter ζ from a value of Γ and its source value Γ₀.

Plumes.ζ2Γ — Function

ζ2Γ(ζ, Γ₀; ΔΓ = 0.0001)

Calculate the plume parameter Γ for a height ζ and source plume parameter Γ₀.

Plumes.LazyPlumeLengthScale — Function

LazyPlumeLengthScale(b₀; α=0.11)

Calculate the distance from the actual source to the virtual source of a lazy plume from the source radius b₀ and entrainment coefficient α.

Plumes.LazyVirtualSourcePosition — Function

LazyVirtualSourcePosition(Γ₀; N=10)

Calculate the virtual source position of a lazy plume with source parameter Γ₀ including N terms in the series approximation.

Plumes.LSFitPlumeW — Function

LSFitPlumeW(z, w, ρₐ; p0 = [0.1, 0.01, 0.999*ρₐ], α=0.11, g = 9.80665)

Least squares fit a theoretical plume velocity curve to w points at z in an ambient with density ρₐ.

The fit returns best fit coefficients [b₀, w₀, ρ₀] and p0 is the vector of guesses for the fitting coefficients, αis the entranment coefficient, andg` is the gravitational acceleration.

Plumes.LazyPlumeW — Function

LazyPlumeW(Γ, Γ₀, w₀)

Calculate the centerline vertical velocity of a lazy plume as a function of Γ and its source value Γ₀ and source velocity w₀.

Plumes.LazyPlumeFarFieldW — Function

LazyPlumeFarFieldW(z, Γ₀, w₀, b₀)

Calculate the asymptotic far-field centerline vertical velocity of a lazy plume as a function of ζ, its source value Γ₀, source velocity w₀.

Plumes.PurePlumeW — Function

PurePlumeW(ζ, b₀, w₀)

Calculate the centerline vertical velocity of a pure plume as a function of ζ, its source radius b₀ and source velocity w₀.

Plumes.LazyPlumeRadius — Function

LazyPlumeRadius(Γ, Γ₀, b₀)

Calculate the radius of a lazy plume as a function of Γ, its source value Γ₀, and source radius b₀.

Plumes.LazyPlumeBuoyancy — Function

LazyPlumeBuoyancy(Γ, Γ₀, g_prime₀)

Calculate the buoyancy of a lazy plume as a function of Γ, its source value Γ₀, and source buoyancy g_prime₀.

Plumes.LazyPlumeTracerConcentration — Function

LazyPlumeTracerConcentration(Γ, Γ₀, g_prime₀)

Calculate the tracer concentration of a lazy plume as a function of Γ, its source value Γ₀, and source buoyancy g_prime₀.

Plumes.LazyPlumeVolumeFlux — Function

LazyPlumeVolumeFlux(Γ, Γ₀, Q₀)

Calculate the volume flux of a lazy plume as a function of Γ, its source value Γ₀, and source volume flux Q₀.

Plumes.LazyPlumeMomentumFlux — Function

LazyPlumeMomentumFlux(Γ, Γ₀, M₀)

Calculate the momentum flux of a lazy plume as a function of Γ, its source value Γ₀, and source momentum flux Q₀.

Plumes.EntrainmentRate — Function

EntrainmentRate(Γ, Γ₀, Q₀, b₀; α=0.11)

Unexported

Plumes.Γ_integral — Function

Γ_integral(Γ, Γ₀)

Calculate the integral contained in equation 4.4 of Hunt and Kaye (2005) from a value of Γ and its source value Γ₀.

Stratified MTT Equations

Plumes.mtt! — Function

mtt!(du, u, p, z, κ)

Update the derivatives of the state vector components of the stratified MTT equations for a given plume shape (top-hat or Gaussian).

For the case where N² is a free parameter. ...

Arguments

du: The derivatives of the state vector components (w, b, g').

u: The state vector (w, b, g', N²).

p: The fixed parameters (α).

z: The spatial coordinate.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.mtt_fixedN²! — Function

mtt_fixedN²!(du, u, p, z, κ)

Update the derivatives of the state vector components of the stratified MTT equations for a given plume shape (top-hat or Gaussian).

For the case where N² is a fixed parameter. ...

Arguments

du: The derivatives of the state vector components (w, b, g').

u: The state vector (w, b, g').

p: The fixed parameters (α, N²).

z: The spatial coordinate.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.mtt_free_α! — Function

mtt_free_α!(du, u, p, z, κ)

Update the derivatives of the state vector components of the stratified MTT equations for a given plume shape (top-hat or Gaussian).

For the case where N² and α are free parameter. ...

Arguments

du: The derivatives of the state vector components (w, b, g').

u: The state vector (w, b, g').

p: The fixed parameters (α, N²).

z: The spatial coordinate.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.mttmodel — Function

mttmodel(z, u0, mttfunc, fixed)

Calculate the vertical velocity at z from boundary conditions u0 for a given set of MTT plume equations mttfunc and fixed parameters.

Plumes.mttmodel_fixed_w₀ — Function

mttmodel_fixed_w₀(z, u0, mttfunc, fixed)

Calculate the vertical velocity at z from boundary conditions u0 for a given set of MTT plume equations mttfunc and fixed parameters.

For the case where the initial vertical velocity w₀ is provided.

Plumes.mttmodel_free_α — Function

mttmodel_free_α(z, u0, mttfunc)

Calculate the vertical velocity at z from boundary conditions u0 for a given set of MTT plume equations mttfunc and fixed parameters.

For the case where the entrainment coefficient α is a free parameter.

Plumes.CalculateMTTProfiles — Function

CalculateMTTProfiles(zs, w₀, b₀, g_prime₀, N², α, κ)

Calculate the profiles of w, b, and g' for a plume described with the stratified MTT equations at zs.

w₀, b₀, and g_prime₀ form the boundary conditions that the solver uses. ...

Arguments

zs: The spatial coordinates at which to evaluate the quantities.

w₀: The source vertical velocity (z=0).

b₀: The source radius (z=0).

g_prime₀: The source reduced gravity (z=0).

N²: The buoyancy frequency squared.

α: The entrainment coefficient.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.CalculateLinearDensityProfile — Function

CalculateLinearDensityProfile(z, ρ₀, ρref, N²; g = 9.80665)

Calculate the density at z for a linear denisty profile with squared buoyancy frequency N², initial density ρ(z=0) = ρ₀, and reference density ρref.

Misc

Plumes.MeanCTDVariable — Function

MeanCTDVariable(time, variable, sttimes, endtimes; max_depth=nothing, depth=nothing)

Find the mean and standard deviation of variable during the time periods specified by sttimes and endtimes.

Optionally remove values below a depth threshold max_depth (could actually be any variable, not just depth).