Plume Quantities

These functions define various plume quantities in general.

Non-Dimensional Numbers

Plumes.PlumeRe — Function

PlumeRe(w, b; ν=1.0e-6)

Calculate the plume Reynolds number from the plume velocity w, plume radius b, and kinematic viscosity ν.

Plumes.PlumeRi — Function

PlumeRi(w, b, g_prime)

Calculate the plume Richardson number from the plume velocity w, plume radius b, and reduced gravity g_prime.

PlumeRi(Γ; κ=2, α=0.11)

Calculate the source Richardson number from the source parameter Γ, shape coefficient κ, and entrainment coefficient α.

Plumes.SourceRa — Function

SourceRa(ρₐ, ρ₀, h; ν=1.0e-6, κ=1.0e-9, g = 9.80665)

Calculate the plume Rayleigh number from the ambient density ρₐ, source density ρ₀, depth h, kinematic viscosity ν, and molecular diffusivity κ.

SourceRa(g_prime₀, h; ν=1.0e-6, κ=1.0e-9)

Calculate the source Rayleigh number from the reduced gravity g_prime₀, depth h, kinematic viscosity ν, and molecular diffusivity κ.

Plumes.PlumeParameterΓ — Function

PlumeParameterΓ(w, b, g_prime; α=0.11)

Calculate the plume parameter Γ.

...

Arguments

b: Plume radius

w: Vertical velocity

g_prime: Reduced gravity

α: Entranment coefficient

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.Γ₀Definition — Function

Γ₀Definition(b₀, w₀, ρ₀, ρₐ; α=0.11, g = 9.80665)

Calculate the source plume parameter Γ₀.

...

Arguments

b₀: Source radius

w₀: Source vertical velocity

ρ₀: Souce density

ρₐ: Ambient density

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

α: Entranment coefficient

g: gravitational acceleration

...

Buoyancy Quantities

Plumes.ReducedGravity — Function

ReducedGravity(ρ₀, ρₐ; g = 9.80665)

Calculate the reduced gravity for density ρ₀ and the ambient ρₐ.

Plumes.ρ_from_g_prime — Function

ρ_from_g_prime(g_prime, ρₐ; ρref=ρₐ, g = 9.80665)

Calculate the density ρ from the reduced gravity g_prime and the ambient density ρₐ.

Fluxes

Plumes.PlumeArea — Function

PlumeArea(b₀)

Area of the source of the plume with source radius b₀.

Total Fluxes

Plumes.TotalBuoyancyFlux — Function

TotalBuoyancyFlux(w₀, b₀, g_prime₀, κ)

Calculate the total buoyancy flux from the source velocity w₀, source radius b₀, and source reduced gravity g_prime₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.TotalMomentumFlux — Function

TotalMomentumFlux(w₀, b₀, κ)

Calculate the total momentum flux from the source velocity w₀, source radius b₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.TotalVolumeFlux — Function

TotalVolumeFlux(w₀, b₀)

Calculate the total volume flux from the source velocity w₀, source radius b₀. Volume flux does not depend on plume shape

Plumes.TotalTracerFlux — Function

TotalTracerFlux(w₀, b₀, κ; C₀=1)

Calculate the total tracer flux from the source velocity w₀, source radius b₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Fluxes per Area

Plumes.BuoyancyFluxPerArea — Function

BuoyancyFluxPerArea(w₀, g_prime₀, κ)

Calculate the per area buoyancy flux from the source velocity w₀ and source reduced gravity g_prime₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.MomentumFluxPerArea — Function

MomentumFluxPerArea(w₀, κ)

Calculate the per area momentum flux from the source velocity w₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.

Plumes.VolumeFluxPerArea — Function

VolumeFluxPerArea(w₀)

Calculate the per area volume flux from the source velocity w₀. The volume flux does not depend on plume shape.

Plumes.TracerFluxPerArea — Function

TracerFluxPerArea(w₀, κ; C₀=1)

Calculate the per area tracer flux from the source velocity w₀ and source concentration C₀. κ determines the plume shape, 1 for top-hat, 2 for Gaussian.