Stratified MTT Equations

These functions relate to the analysis and fitting of the general stratified MTT plume equations.

ODEs

Plumes.mtt! — Function

mtt!(du, u, p, z, κ)

Update the derivatives of the state vector components of the stratified MTT equations for a given plume shape (top-hat or Gaussian).

For the case where N² is a free parameter. ...

Arguments

du: The derivatives of the state vector components (w, b, g').

u: The state vector (w, b, g', N²).

p: The fixed parameters (α).

z: The spatial coordinate.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.mtt_fixedN²! — Function

mtt_fixedN²!(du, u, p, z, κ)

Update the derivatives of the state vector components of the stratified MTT equations for a given plume shape (top-hat or Gaussian).

For the case where N² is a fixed parameter. ...

Arguments

du: The derivatives of the state vector components (w, b, g').

u: The state vector (w, b, g').

p: The fixed parameters (α, N²).

z: The spatial coordinate.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.mtt_free_α! — Function

mtt_free_α!(du, u, p, z, κ)

Update the derivatives of the state vector components of the stratified MTT equations for a given plume shape (top-hat or Gaussian).

For the case where N² and α are free parameter. ...

Arguments

du: The derivatives of the state vector components (w, b, g').

u: The state vector (w, b, g').

p: The fixed parameters (α, N²).

z: The spatial coordinate.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Fit Models

Plumes.mttmodel — Function

mttmodel(z, u0, mttfunc, fixed)

Calculate the vertical velocity at z from boundary conditions u0 for a given set of MTT plume equations mttfunc and fixed parameters.

Plumes.mttmodel_fixed_w₀ — Function

mttmodel_fixed_w₀(z, u0, mttfunc, fixed)

Calculate the vertical velocity at z from boundary conditions u0 for a given set of MTT plume equations mttfunc and fixed parameters.

For the case where the initial vertical velocity w₀ is provided.

Plumes.mttmodel_free_α — Function

mttmodel_free_α(z, u0, mttfunc)

Calculate the vertical velocity at z from boundary conditions u0 for a given set of MTT plume equations mttfunc and fixed parameters.

For the case where the entrainment coefficient α is a free parameter.

Plume Quantity Profiles

Plumes.CalculateMTTProfiles — Function

CalculateMTTProfiles(zs, w₀, b₀, g_prime₀, N², α, κ)

Calculate the profiles of w, b, and g' for a plume described with the stratified MTT equations at zs.

w₀, b₀, and g_prime₀ form the boundary conditions that the solver uses. ...

Arguments

zs: The spatial coordinates at which to evaluate the quantities.

w₀: The source vertical velocity (z=0).

b₀: The source radius (z=0).

g_prime₀: The source reduced gravity (z=0).

N²: The buoyancy frequency squared.

α: The entrainment coefficient.

κ: Factor determining plume shape. 1 for top-hat (circle), 2 for Gaussian.

...

Plumes.CalculateLinearDensityProfile — Function

CalculateLinearDensityProfile(z, ρ₀, ρref, N²; g = 9.80665)

Calculate the density at z for a linear denisty profile with squared buoyancy frequency N², initial density ρ(z=0) = ρ₀, and reference density ρref.