Unstratified Lazy Plume Equations

These functions relate to the analysis and fitting of unstratified lazy plumes.

Coordinate Transforms

Plumes.z2ζ — Function

z2ζ(z, b₀; α=0.11)

Convert depth z to scaled plume height ζ using the source parameter b₀ and the entrainment coefficient α.

Plumes.ζ2z — Function

ζ2z(ζ, b₀; α=0.11)

Convert scaled plume height ζ to depth z using the source parameter b₀ and the entrainment coefficient α.

Plumes.Γ2ζ — Function

Γ2ζ(Γ, Γ₀)

Calculate the height parameter ζ from a value of Γ and its source value Γ₀.

Plumes.ζ2Γ — Function

ζ2Γ(ζ, Γ₀; ΔΓ = 0.0001)

Calculate the plume parameter Γ for a height ζ and source plume parameter Γ₀.

Lengthscales

Plumes.LazyPlumeLengthScale — Function

LazyPlumeLengthScale(b₀; α=0.11)

Calculate the distance from the actual source to the virtual source of a lazy plume from the source radius b₀ and entrainment coefficient α.

Plumes.LazyVirtualSourcePosition — Function

LazyVirtualSourcePosition(Γ₀; N=10)

Calculate the virtual source position of a lazy plume with source parameter Γ₀ including N terms in the series approximation.

Fitting

Plumes.LSFitPlumeW — Function

LSFitPlumeW(z, w, ρₐ; p0 = [0.1, 0.01, 0.999*ρₐ], α=0.11, g = 9.80665)

Least squares fit a theoretical plume velocity curve to w points at z in an ambient with density ρₐ.

The fit returns best fit coefficients [b₀, w₀, ρ₀] and p0 is the vector of guesses for the fitting coefficients, αis the entranment coefficient, andg` is the gravitational acceleration.

Plume Quantitiy Profiles

Plumes.LazyPlumeW — Function

LazyPlumeW(Γ, Γ₀, w₀)

Calculate the centerline vertical velocity of a lazy plume as a function of Γ and its source value Γ₀ and source velocity w₀.

Plumes.LazyPlumeFarFieldW — Function

LazyPlumeFarFieldW(z, Γ₀, w₀, b₀)

Calculate the asymptotic far-field centerline vertical velocity of a lazy plume as a function of ζ, its source value Γ₀, source velocity w₀.

Plumes.PurePlumeW — Function

PurePlumeW(ζ, b₀, w₀)

Calculate the centerline vertical velocity of a pure plume as a function of ζ, its source radius b₀ and source velocity w₀.

Plumes.LazyPlumeRadius — Function

LazyPlumeRadius(Γ, Γ₀, b₀)

Calculate the radius of a lazy plume as a function of Γ, its source value Γ₀, and source radius b₀.

Plumes.LazyPlumeBuoyancy — Function

LazyPlumeBuoyancy(Γ, Γ₀, g_prime₀)

Calculate the buoyancy of a lazy plume as a function of Γ, its source value Γ₀, and source buoyancy g_prime₀.

Plumes.LazyPlumeTracerConcentration — Function

LazyPlumeTracerConcentration(Γ, Γ₀, g_prime₀)

Calculate the tracer concentration of a lazy plume as a function of Γ, its source value Γ₀, and source buoyancy g_prime₀.

Plumes.LazyPlumeVolumeFlux — Function

LazyPlumeVolumeFlux(Γ, Γ₀, Q₀)

Calculate the volume flux of a lazy plume as a function of Γ, its source value Γ₀, and source volume flux Q₀.

Plumes.LazyPlumeMomentumFlux — Function

LazyPlumeMomentumFlux(Γ, Γ₀, M₀)

Calculate the momentum flux of a lazy plume as a function of Γ, its source value Γ₀, and source momentum flux Q₀.

Plumes.EntrainmentRate — Function

EntrainmentRate(Γ, Γ₀, Q₀, b₀; α=0.11)